implement channel txnID restore
The `restoreTxnID` is used to set the txnID marker when we restore
from shutdown (or a Channel is reopened after being shutdown, once
that's implemented).

Note: the reverse iterator for badger is screamingly unintuitive and
hard to get right just from reading the documentation; if you're
editing this code make sure the semantics here are preserved!