initial implementation of Store subscriptions
The Store serves as an intermediary between the Server and the p2p
host, with both components acting as subscribers to channels, which
have been implemented as key prefixes.

Badger DB has snapshot isolation for all views, so we can't just
continuously iterate to pick up new entries, and the badger `Subscribe`
API can only get us new entries and not historical ones.

Instead, we'll iterate over a snapshot and the subscription's channel
will notify us if any new entries have been written, which tells us
that the iterator is "dirty" and we can start a new iterator once we
reach the end if we still want more entries.