update design notes for LSM key schema
EntryIDs are 30 bytes, sortable by timestamp and including a random
component sufficient to prevent collisions.

Because peers can come and go, we may get "historical" entries from
those peers from a large time window in the past (configurable by the
community but expected to be on the order of days and weeks, not
seconds).

Therefore, in addition to the Entries keyspace, we'll have an
ArrivalIndex keyspace (with zero-values) which gives us ordered keys
for when entries were received *by this peer*. A client that comes
online can use its last ArrivalIndex to ask the vault for the keys.