store: enforce append-only semantics
Duplicate keys should return an error when we append. Badger can't
give us true "append-only", but we can check for the key's existence
in the write transaction.

Note: we can't guarantee ordering in case of a tight race because the
check happens within a snapshot. Clients should be verifying the
content hash portion of the key before accepting entries. That limits
any attacker to DoS'ing writes at worst, assuming they can observe
entry keys in transit. This is protected against by TLS between the
client and vault.