server: implement feed close and add closing contexts
Split out the differences between a Stream, which is tracked by
client-controller ReqID, and the Connection which is tracked by
internal UUID. This allows clients to close streams they're no longer
interested in without closing the whole connection.

Thread a context down from main into the subscribers so we can
gracefully shutdown.