hack: prevent test deadlock because of dropped notification
The OpenFeed request in `TestServer_NewOpenClose` may not get a chance
to set up the badger Subscribe before we send the append entry, so we
can end up dropping the notification. For now we'll just sleep for
this test and come back later to see if we can make the Subscribe
synchronous.